Combine repeated definition of Usage model into one
